Made some minor fixes and improvements: - Tab order on Register Form is now fixed - Being on Password textbox or the Admin button, the user can now press enter on the keyboard and it will create the user. - The Register form will now clear when the user presses Register or presses the enter key. - Tester Report Bug form now has a dropdown list to select a Tester instead of an input field
Unfortunately, it seems that the database path needed changing again as it no longer exists in the Debug folder. I must investigate later to see if there is a way to avoid having to change the SqlConnection path every time I am on a new computer.
Also to do is looking into Unit Tests and integrating the program with Source Control.
For now, I will work on getting the View Bug form to work as intended.